+++ src/main/java/com/shortener/util/NumberSystem.java
@@ -1,5 +1,5 @@
 package com.shortener.util;
 
 public interface NumberSystem {
-   String getConvertedValue(int base10Number);
+   String getConvertedValue(long base10Number);
 }
